+++ Wrappers/Python/cil/optimisation/algorithms/SPDHG.py
@@ -22,7 +22,7 @@ from cil.optimisation.algorithms import Algorithm
 import numpy as np
 import warnings
 import logging
-from cil.optimisation.algorithms import Sampler
+from cil.optimisation.utilities import Sampler
 from numbers import Number
